feat: merge Command and Script
Really only Script is needed, as it generalizes Command.

This also makes it so instead of running each line of a script in
a separate Bash process, we just pass the entire script to one Bash
invocation. So now we support real Bash scripting, e.g. variables and
control flow.

Lois thought there was a reason we were doing the line-splitting stuff,
but couldn't remember why. I figure we should try this change and see
if anything blows up. We can always revert it pretty easily.