fix: dont tag buildcache, but change cache name
Currently our cache is a tag, so building v2 of myimage would push

- myimage:v2
- myimage:latest
- myimage:buildcache

The buildcache one cannot be deployed, but could get fetch by updatecli

This changes the image to be

- myimage:v2
- myimage:latest
- myimage-buildcache

Where `myimage-buildcache` is an entirely different image

We use updatecli to fetch the latest version of the images.

Sometimes, it finds buildcache as the latest version. It's not true, it's the latest cache version

So i change the cache to be another image instead of another version of the base image