feat: improve default args for fix-lock-files
Previously, the default behavior of fix-lock-files was based on
DiffStrategy::LocalChanges. This behavior was not useful in the most
common case of a developer trying to fix their PR to pass CI.
Many developers rightfully got confused by this.

Now the default behavior is to use "origin/main" as the base revspec,
which is correct for most PRs. I made sure to add enough docs in the CLI
help so developers know what to do if they're doing something less common,
like a backport PR.

For extra clarity, I decoupled the fix-lock-files command from the
DiffOptions and DiffStrategy types. DiffStrategy carries semantics that
fix-lock-files does not care about, like how to select which packages
need to be tested or published.